+++ ktransformers/ktransformers_ext/operators/llamafile/moe.h
@@ -51,9 +51,9 @@ class MOE {
     MOE(MOEConfig);
     ~MOE();
     void warm_up(Backend* backend);
-    void forward_one(int k, const uint64_t* expert_ids, const float* weights, const void* input, void* output, Backend* backend);
-    void forward_many(int qlen, int k, const uint64_t* expert_ids, const float* weights, const void* input, void* output, Backend* backend);
-    void forward(int qlen, int k, const uint64_t* expert_ids, const float* weights, const void* input, void* output, Backend* backend);
+    void forward_one(int k, const uint32_t* expert_ids, const float* weights, const void* input, void* output, Backend* backend);
+    void forward_many(int qlen, int k, const uint32_t* expert_ids, const float* weights, const void* input, void* output, Backend* backend);
+    void forward(int qlen, int k, const uint32_t* expert_ids, const float* weights, const void* input, void* output, Backend* backend);
 
    private:
     MOEConfig config_;
